Tailwind CSS and updated README
Using the Markdown All in One extension for building the README.

Greatly improved the README by also including the extensions used.

Updated the .vscode settings to apply for the project, regardless of the
user settings. Also, added the support with emmet to allow
autocompletion.

Added the rule to .stylelintrc to add a space after a rule declaration
semicolon.

Modified the .stylelintrc to ignore the font-family quotes.