Updated workflows and package.json
Removed the optionalDependencies for fsevents

Improved the build workflow, formerly build_and_test, so that it caches
npm dependencies installed. It also has seperate jobs for the linting,
testing, and the build.

Improved the deploy workflow to essentially function the same as the
build workflow but, the deploy depends on the build job.